rename the service file
