+++ packages/client/hmi-client/src/components/model/petrinet/tera-model-equation.vue
@@ -31,7 +31,7 @@ import {
 	convertAMRToACSet,
 	updateExistingModelContent
 } from '@/model-representation/petrinet/petrinet-service';
-import { latexToAMR } from '@/services/extractions';
+import { latexToAMR } from '@/services/knowledge';
 import { cleanLatexEquations } from '@/utils/math';
 import { petriToLatex } from '@/petrinet/petrinet-service';
 import { isEmpty } from 'lodash';
